Sync Fork from Upstream Repo

+++ src/main/java/com/google/devtools/build/lib/analysis/test/TestRunnerAction.java
@@ -16,6 +16,7 @@ package com.google.devtools.build.lib.analysis.test;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
+import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
@@ -33,10 +34,10 @@ import com.google.devtools.build.lib.actions.ActionOwner;
 import com.google.devtools.build.lib.actions.ActionResult;
 import com.google.devtools.build.lib.actions.Artifact;
 import com.google.devtools.build.lib.actions.ArtifactPathResolver;
+import com.google.devtools.build.lib.actions.CommandAction;
 import com.google.devtools.build.lib.actions.CommandLineExpansionException;
 import com.google.devtools.build.lib.actions.EnvironmentalExecException;
 import com.google.devtools.build.lib.actions.ExecException;
-import com.google.devtools.build.lib.actions.ExecutionInfoSpecifier;
 import com.google.devtools.build.lib.actions.NotifyOnActionCacheHit;
 import com.google.devtools.build.lib.actions.RunfilesSupplier;
 import com.google.devtools.build.lib.actions.SpawnExecutedEvent;
@@ -51,6 +52,10 @@ import com.google.devtools.build.lib.analysis.test.TestActionContext.TestRunnerS
 import com.google.devtools.build.lib.buildeventstream.TestFileNameConstants;
 import com.google.devtools.build.lib.cmdline.Label;
 import com.google.devtools.build.lib.collect.ImmutableIterable;
+import com.google.devtools.build.lib.collect.nestedset.NestedSet;
+import com.google.devtools.build.lib.collect.nestedset.NestedSetBuilder;
+import com.google.devtools.build.lib.collect.nestedset.Order;
+import com.google.devtools.build.lib.exec.TestStrategy;
 import com.google.devtools.build.lib.util.Fingerprint;
 import com.google.devtools.build.lib.util.LoggingUtil;
 import com.google.devtools.build.lib.util.Pair;
@@ -59,6 +64,7 @@ import com.google.devtools.build.lib.vfs.Path;
 import com.google.devtools.build.lib.vfs.PathFragment;
 import com.google.devtools.build.lib.view.test.TestStatus.TestResultData;
 import com.google.devtools.common.options.TriState;
+import com.google.protobuf.ExtensionRegistry;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -76,7 +82,7 @@ import javax.annotation.Nullable;
  */
 // Not final so that we can mock it in tests.
 public class TestRunnerAction extends AbstractAction
-    implements NotifyOnActionCacheHit, ExecutionInfoSpecifier {
+    implements NotifyOnActionCacheHit, CommandAction {
   public static final PathFragment COVERAGE_TMP_ROOT = PathFragment.create("_coverage");
 
   // Used for selecting subset of testcase / testmethods.
@@ -109,14 +115,12 @@ public class TestRunnerAction extends AbstractAction
   private final PathFragment testInfrastructureFailure;
   private final PathFragment baseDir;
   private final Artifact coverageData;
+  @Nullable private final Artifact coverageDirectory;
   private final TestTargetProperties testProperties;
   private final TestTargetExecutionSettings executionSettings;
   private final int shardNum;
   private final int runNumber;
   private final String workspaceName;
-
-  // Takes the value of the `--windows_native_test_wrapper` flag.
-  private final boolean useTestWrapperInsteadOfTestSetupSh;
 
   // Mutable state related to test caching. Lazily initialized: null indicates unknown.
   private Boolean unconditionalExecution;
@@ -132,8 +136,8 @@ public class TestRunnerAction extends AbstractAction
 
   private final boolean cancelConcurrentTestsOnSuccess;
 
-  private static ImmutableList<Artifact> list(Artifact... artifacts) {
-    ImmutableList.Builder<Artifact> builder = ImmutableList.builder();
+  private static ImmutableSet<Artifact> nonNullAsSet(Artifact... artifacts) {
+    ImmutableSet.Builder<Artifact> builder = ImmutableSet.builder();
     for (Artifact artifact : artifacts) {
       if (artifact != null) {
         builder.add(artifact);
@@ -152,15 +156,15 @@ public class TestRunnerAction extends AbstractAction
    */
   TestRunnerAction(
       ActionOwner owner,
-      Iterable<Artifact> inputs,
+      NestedSet<Artifact> inputs,
       RunfilesSupplier runfilesSupplier,
       Artifact testSetupScript, // Must be in inputs
-      boolean useTestWrapperInsteadOfTestSetupSh,
       Artifact testXmlGeneratorScript, // Must be in inputs
       @Nullable Artifact collectCoverageScript, // Must be in inputs, if not null
       Artifact testLog,
       Artifact cacheStatus,
       Artifact coverageArtifact,
+      @Nullable Artifact coverageDirectory,
       TestTargetProperties testProperties,
       Map<String, String> extraTestEnv,
       TestTargetExecutionSettings executionSettings,
@@ -169,17 +173,17 @@ public class TestRunnerAction extends AbstractAction
       BuildConfiguration configuration,
       String workspaceName,
       @Nullable PathFragment shExecutable,
-      boolean cancelConcurrentTestsOnSuccess) {
+      boolean cancelConcurrentTestsOnSuccess,
+      Iterable<Artifact> tools) {
     super(
         owner,
-        /*tools=*/ ImmutableList.of(),
+        NestedSetBuilder.wrap(Order.STABLE_ORDER, tools),
         inputs,
         runfilesSupplier,
-        list(testLog, cacheStatus, coverageArtifact),
+        nonNullAsSet(testLog, cacheStatus, coverageArtifact, coverageDirectory),
         configuration.getActionEnvironment());
     Preconditions.checkState((collectCoverageScript == null) == (coverageArtifact == null));
     this.testSetupScript = testSetupScript;
-    this.useTestWrapperInsteadOfTestSetupSh = useTestWrapperInsteadOfTestSetupSh;
     this.testXmlGeneratorScript = testXmlGeneratorScript;
     this.collectCoverageScript = collectCoverageScript;
     this.configuration = Preconditions.checkNotNull(configuration);
@@ -188,6 +192,7 @@ public class TestRunnerAction extends AbstractAction
     this.testLog = testLog;
     this.cacheStatus = cacheStatus;
     this.coverageData = coverageArtifact;
+    this.coverageDirectory = coverageDirectory;
     this.shardNum = shardNum;
     this.runNumber = runNumber;
     this.testProperties = Preconditions.checkNotNull(testProperties);
@@ -255,7 +260,10 @@ public class TestRunnerAction extends AbstractAction
     outputs.add(ActionInputHelper.fromPath(getUndeclaredOutputsManifestPath()));
     outputs.add(ActionInputHelper.fromPath(getUndeclaredOutputsAnnotationsPath()));
     if (isCoverageMode()) {
-      outputs.add(getCoverageData());
+      outputs.add(coverageData);
+      if (coverageDirectory != null) {
+        outputs.add(coverageDirectory);
+      }
     }
     return outputs;
   }
@@ -328,8 +336,8 @@ public class TestRunnerAction extends AbstractAction
       throws CommandLineExpansionException {
     fp.addString(GUID);
     fp.addStrings(executionSettings.getArgs().arguments());
-    fp.addString(
-        executionSettings.getTestFilter() == null ? "" : executionSettings.getTestFilter());
+    fp.addString(Strings.nullToEmpty(executionSettings.getTestFilter()));
+    fp.addBoolean(executionSettings.getTestRunnerFailFast());
     RunUnder runUnder = executionSettings.getRunUnder();
     fp.addString(runUnder == null ? "" : runUnder.getValue());
     fp.addStringMap(extraTestEnv);
@@ -360,6 +368,7 @@ public class TestRunnerAction extends AbstractAction
     return unconditionalExecution;
   }
 
+
   @Override
   public boolean isVolatile() {
     return true;
@@ -377,7 +386,7 @@ public class TestRunnerAction extends AbstractAction
   @Nullable
   private TestResultData readCacheStatus() {
     try (InputStream in = cacheStatus.getPath().getInputStream()) {
-      return TestResultData.parseFrom(in);
+      return TestResultData.parseFrom(in, ExtensionRegistry.getEmptyRegistry());
     } catch (IOException expected) {
       return null;
     }
@@ -484,7 +493,8 @@ public class TestRunnerAction extends AbstractAction
     deleteTestAttemptsDirMaybe(execRoot.getRelative(baseDir), "test");
   }
 
-  private void deleteTestAttemptsDirMaybe(Path outputDir, String namePrefix) throws IOException {
+  private static void deleteTestAttemptsDirMaybe(Path outputDir, String namePrefix)
+      throws IOException {
     Path testAttemptsDir = outputDir.getChild(namePrefix + "_attempts");
     if (testAttemptsDir.exists()) {
       // Normally we should have used deleteTree(testAttemptsDir). However, if test output is
@@ -498,6 +508,12 @@ public class TestRunnerAction extends AbstractAction
       } catch (IOException e) {
         // Do nothing.
       }
+    }
+  }
+
+  void createEmptyOutputs(ActionExecutionContext context) throws IOException {
+    for (Artifact output : TestRunnerAction.this.getMandatoryOutputs()) {
+      FileSystemUtils.touchFile(context.getInputPath(output));
     }
   }
 
@@ -519,6 +535,9 @@ public class TestRunnerAction extends AbstractAction
     String testFilter = getExecutionSettings().getTestFilter();
     if (testFilter != null) {
       env.put(TEST_BRIDGE_TEST_FILTER_ENV, testFilter);
+    }
+    if (testConfiguration.getTestRunnerFailFast()) {
+      env.put("TESTBRIDGE_TEST_RUNNER_FAIL_FAST", "1");
     }
 
     env.put("TEST_WARNINGS_OUTPUT_FILE", getTestWarningsPath().getPathString());
@@ -549,6 +568,11 @@ public class TestRunnerAction extends AbstractAction
     if (!isEnableRunfiles()) {
       // If runfiles are disabled, tell remote-runtest.sh/local-runtest.sh about that.
       env.put("RUNFILES_MANIFEST_ONLY", "1");
+    }
+
+    if (testConfiguration.isPersistentTestRunner()) {
+      // Let the test runner know it runs persistently within a worker.
+      env.put("PERSISTENT_TEST_RUNNER", "true");
     }
 
     if (isCoverageMode()) {
@@ -630,12 +654,12 @@ public class TestRunnerAction extends AbstractAction
     return undeclaredOutputsDir;
   }
 
-  /** @return path to the optional zip file of undeclared test outputs. */
+  /** Returns path to the optional zip file of undeclared test outputs. */
   public PathFragment getUndeclaredOutputsZipPath() {
     return undeclaredOutputsZipPath;
   }
 
-  /** @return path to the undeclared output manifest file. */
+  /** Returns path to the undeclared output manifest file. */
   public PathFragment getUndeclaredOutputsManifestPath() {
     return undeclaredOutputsManifestPath;
   }
@@ -644,7 +668,7 @@ public class TestRunnerAction extends AbstractAction
     return undeclaredOutputsAnnotationsDir;
   }
 
-  /** @return path to the undeclared output annotations file. */
+  /** Returns path to the undeclared output annotations file. */
   public PathFragment getUndeclaredOutputsAnnotationsPath() {
     return undeclaredOutputsAnnotationsPath;
   }
@@ -661,12 +685,12 @@ public class TestRunnerAction extends AbstractAction
     return testInfrastructureFailure;
   }
 
-  /** @return path to the optionally created XML output file created by the test. */
+  /** Returns path to the optionally created XML output file created by the test. */
   public PathFragment getXmlOutputPath() {
     return xmlOutputPath;
   }
 
-  /** @return coverage data artifact or null if code coverage was not requested. */
+  /** Returns coverage data artifact or null if code coverage was not requested. */
   @Nullable
   public Artifact getCoverageData() {
     return coverageData;
@@ -688,11 +712,18 @@ public class TestRunnerAction extends AbstractAction
    * execution with exception of the locally generated *.gcda files. Those are stored separately
    * using relative path within coverage directory.
    *
-   * <p>The directory name for the given test runner action is constructed as: {@code
+   * <p>If the coverageDirectory field is set, then its exec path is returned. This is a tree
+   * artifact, meaning that all files in the corresponding directories are returned from sandboxed
+   * or remote execution.
+   *
+   * <p>Otherwise, the directory name for the given test runner action is constructed as: {@code
    * _coverage/target_path/test_log_name} where {@code test_log_name} is usually a target name but
    * potentially can include extra suffix, such as a shard number (if test execution was sharded).
    */
   public PathFragment getCoverageDirectory() {
+    if (coverageDirectory != null) {
+      return coverageDirectory.getExecPath();
+    }
     return COVERAGE_TMP_ROOT.getRelative(
         FileSystemUtils.removeExtension(getTestLog().getRootRelativePath()));
   }
@@ -715,19 +746,19 @@ public class TestRunnerAction extends AbstractAction
   }
 
   /**
-   * @return the shard number for this action. If getTotalShards() > 0, must be >= 0 and <
-   *     getTotalShards(). Otherwise, must be 0.
+   * Returns the shard number for this action. If getTotalShards() > 0, must be >= 0 and <
+   * getTotalShards(). Otherwise, must be 0.
    */
   public int getShardNum() {
     return shardNum;
   }
 
-  /** @return run number. */
+  /** Returns run number. */
   public int getRunNumber() {
     return runNumber;
   }
 
-  /** @return the workspace name. */
+  /** Returns the workspace name. */
   public String getRunfilesPrefix() {
     return workspaceName;
   }
@@ -760,6 +791,8 @@ public class TestRunnerAction extends AbstractAction
           beginIfNotCancelled(testRunnerSpawn, cancelFuture);
       if (testAttemptContinuation == null) {
         testRunnerSpawn.finalizeCancelledTest(ImmutableList.of());
+        // We need to create the mandatory output files even if we're not going to run anything.
+        createEmptyOutputs(actionExecutionContext);
         return ActionContinuationOrResult.of(ActionResult.create(ImmutableList.of()));
       }
       return new RunAttemptsContinuation(
@@ -835,10 +868,6 @@ public class TestRunnerAction extends AbstractAction
     return testSetupScript;
   }
 
-  public boolean isUsingTestWrapperInsteadOfTestSetupScript() {
-    return useTestWrapperInsteadOfTestSetupSh;
-  }
-
   public Artifact getTestXmlGeneratorScript() {
     return testXmlGeneratorScript;
   }
@@ -859,6 +888,22 @@ public class TestRunnerAction extends AbstractAction
 
   public boolean isEnableRunfiles() {
     return configuration.runfilesEnabled();
+  }
+
+  @Override
+  public List<String> getArguments() throws CommandLineExpansionException {
+    return TestStrategy.expandedArgsFromAction(this);
+  }
+
+  @Override
+  public ImmutableMap<String, String> getIncompleteEnvironmentForTesting()
+      throws ActionExecutionException {
+    return getEnvironment().getFixedEnv().toMap();
+  }
+
+  @Override
+  public NestedSet<Artifact> getPossibleInputsForTesting() {
+    return getInputs();
   }
 
   /** The same set of paths as the parent test action, resolved against a given exec root. */
@@ -897,32 +942,32 @@ public class TestRunnerAction extends AbstractAction
       return getPath(unusedRunfilesLogPath);
     }
 
-    /** @return path to the directory containing the split logs (raw and proto file). */
+    /** Returns path to the directory containing the split logs (raw and proto file). */
     public Path getSplitLogsDir() {
       return getPath(splitLogsDir);
     }
 
-    /** @return path to the optional zip file of undeclared test outputs. */
+    /** Returns path to the optional zip file of undeclared test outputs. */
     public Path getUndeclaredOutputsZipPath() {
       return getPath(undeclaredOutputsZipPath);
     }
 
-    /** @return path to the directory to hold undeclared test outputs. */
+    /** Returns path to the directory to hold undeclared test outputs. */
     public Path getUndeclaredOutputsDir() {
       return getPath(undeclaredOutputsDir);
     }
 
-    /** @return path to the directory to hold undeclared output annotations parts. */
+    /** Returns path to the directory to hold undeclared output annotations parts. */
     public Path getUndeclaredOutputsAnnotationsDir() {
       return getPath(undeclaredOutputsAnnotationsDir);
     }
 
-    /** @return path to the undeclared output manifest file. */
+    /** Returns path to the undeclared output manifest file. */
     public Path getUndeclaredOutputsManifestPath() {
       return getPath(undeclaredOutputsManifestPath);
     }
 
-    /** @return path to the undeclared output annotations file. */
+    /** Returns path to the undeclared output annotations file. */
     public Path getUndeclaredOutputsAnnotationsPath() {
       return getPath(undeclaredOutputsAnnotationsPath);
     }
@@ -940,7 +985,7 @@ public class TestRunnerAction extends AbstractAction
       return getPath(testInfrastructureFailure);
     }
 
-    /** @return path to the optionally created XML output file created by the test. */
+    /** Returns path to the optionally created XML output file created by the test. */
     public Path getXmlOutputPath() {
       return getPath(xmlOutputPath);
     }
@@ -1023,9 +1068,7 @@ public class TestRunnerAction extends AbstractAction
           if (cancelFuture != null && cancelFuture.isCancelled()) {
             // Clear the interrupt bit.
             Thread.interrupted();
-            for (Artifact output : TestRunnerAction.this.getMandatoryOutputs()) {
-              FileSystemUtils.touchFile(output.getPath());
-            }
+            createEmptyOutputs(testRunnerSpawn.getActionExecutionContext());
             testRunnerSpawn.finalizeCancelledTest(failedAttempts);
             return ActionContinuationOrResult.of(ActionResult.create(spawnResults));
           }
@@ -1057,12 +1100,14 @@ public class TestRunnerAction extends AbstractAction
     private ActionContinuationOrResult process(TestAttemptResult result, int actualMaxAttempts)
         throws ExecException, IOException, InterruptedException {
       spawnResults.addAll(result.spawnResults());
-      if (result.hasPassed()) {
+      TestAttemptResult.Result testResult = result.result();
+      if (testResult == TestAttemptResult.Result.PASSED) {
         if (cancelFuture != null) {
           cancelFuture.cancel(true);
         }
       } else {
-        boolean runAnotherAttempt = failedAttempts.size() + 1 < actualMaxAttempts;
+        boolean runAnotherAttempt =
+            testResult.canRetry() && failedAttempts.size() + 1 < actualMaxAttempts;
         TestRunnerSpawn nextRunner;
         if (runAnotherAttempt) {
           nextRunner = testRunnerSpawn;
@@ -1081,6 +1126,8 @@ public class TestRunnerAction extends AbstractAction
           TestAttemptContinuation nextContinuation = beginIfNotCancelled(nextRunner, cancelFuture);
           if (nextContinuation == null) {
             testRunnerSpawn.finalizeCancelledTest(failedAttempts);
+            // We need to create the mandatory output files even if we're not going to run anything.
+            createEmptyOutputs(testRunnerSpawn.getActionExecutionContext());
             return ActionContinuationOrResult.of(ActionResult.create(spawnResults));
           }
 
@@ -1102,7 +1149,7 @@ public class TestRunnerAction extends AbstractAction
       }
       testRunnerSpawn.finalizeTest(result, failedAttempts);
 
-      if (!keepGoing && !result.hasPassed()) {
+      if (!keepGoing && testResult != TestAttemptResult.Result.PASSED) {
         throw new TestExecException("Test failed: aborting");
       }
       return ActionContinuationOrResult.of(ActionResult.create(spawnResults));
